Decouple snapshot adapter from pg listener
