+++ pg2pg.env
@@ -1,5 +1,6 @@
 # Listener config
 PGSTREAM_POSTGRES_LISTENER_URL="postgres://postgres:postgres@localhost?sslmode=disable"
+PGSTREAM_POSTGRES_LISTENER_INITIAL_SNAPSHOT_ENABLED=true
 PGSTREAM_POSTGRES_INITIAL_SNAPSHOT_TABLES="*"
 PGSTREAM_POSTGRES_INITIAL_SNAPSHOT_SCHEMA_WORKERS=4
 PGSTREAM_POSTGRES_INITIAL_SNAPSHOT_TABLE_WORKERS=4
