+++ pkg/wal/listener/snapshot/adapter/config_test.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
-package postgres
+package adapter
 
 import (
 	"testing"
