+++ pkg/wal/listener/snapshot/wal_snapshot_listener.go
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: Apache-2.0
+
+package snapshot
+
+import "context"
+
+type Generator interface {
+	CreateSnapshot(context.Context) error
+	Close() error
+}
